First rough implementation of onFollow
Implementation of onFollow by polling the API.

Issue realized: The API only allows for 30 requests per minute but for
each channel monitored, one request is required in addition to a few
more requests for converting between username and id. Sending a Bearer
token will increase the limit to 60.

Testing required: Using the chat token for the API to increase limit?

Missing: Implementation of the preparation in case a new Follow-Poll is
created -> Request the followers of the channel and store the most
recent one in map

+++ src/com/cavariux/twitchirc/TwitchAPIConnections/ChannelQueries.java
@@ -16,8 +16,8 @@ public class ChannelQueries {
 	private static final String V5_API_BASE = "https://api.twitch.tv/kraken/channels/";
 
 	public static boolean isSubscriber(Channel channel, User user, String oAuthToken, String clientId) {
-		int[] ids = UserQueries.getUserId(clientId, user.toString(), channel.toString().substring(1));
-		int userId = ids[0], channelId = ids[1];
+		long[] ids = UserQueries.getUserId(clientId, user.toString(), channel.toString().substring(1));
+		long userId = ids[0], channelId = ids[1];
 		String requestUri = V5_API_BASE + channelId + "/subscriptions/" + userId;
 		
 		try {
